profile was divided into profile_info and profile_achievements
